requirements.txt
# Core dependencies
curl-cffi>=0.5.9
pydantic>=2.0.0
python-dateutil>=2.8.2

# Async support
aiofiles>=23.0.0

# JavaScript execution for signature generation
PyExecJS>=1.5.1

# Logging
structlog>=23.0.0

# Type checking
typing-extensions>=4.0.0; python_version < '3.10'